Step4 - 요금 조회

+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -1,15 +1,20 @@
 package nextstep.subway.member.ui;
 
-import nextstep.subway.auth.domain.LoginMember;
+import java.net.URI;
+import nextstep.subway.auth.application.InvalidTokenException;
 import nextstep.subway.auth.domain.AuthenticationPrincipal;
 import nextstep.subway.auth.domain.LoginMember;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.dto.MemberRequest;
 import nextstep.subway.member.dto.MemberResponse;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-import java.net.URI;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 public class MemberController {
@@ -46,6 +51,10 @@ public class MemberController {
     @GetMapping("/members/me")
     public ResponseEntity<MemberResponse> findMemberOfMine(
         @AuthenticationPrincipal LoginMember loginMember) {
+        if (loginMember.isGuest()) {
+            throw new InvalidTokenException("로그인을 하지 않았습니다.");
+        }
+
         MemberResponse member = memberService.findMember(loginMember.getId());
         return ResponseEntity.ok().body(member);
     }
@@ -53,6 +62,10 @@ public class MemberController {
     @PutMapping("/members/me")
     public ResponseEntity<MemberResponse> updateMemberOfMine(
         @AuthenticationPrincipal LoginMember loginMember, @RequestBody MemberRequest param) {
+        if (loginMember.isGuest()) {
+            throw new InvalidTokenException("로그인을 하지 않았습니다.");
+        }
+
         memberService.updateMember(loginMember.getId(), param);
         return ResponseEntity.ok().build();
     }
@@ -60,6 +73,10 @@ public class MemberController {
     @DeleteMapping("/members/me")
     public ResponseEntity<MemberResponse> deleteMemberOfMine(
         @AuthenticationPrincipal LoginMember loginMember) {
+        if (loginMember.isGuest()) {
+            throw new InvalidTokenException("로그인을 하지 않았습니다.");
+        }
+
         memberService.deleteMember(loginMember.getId());
         return ResponseEntity.noContent().build();
     }
